# Optimize scalar mult