--- a/internal/edwards448/scalarmult_test.go
+++ b/internal/edwards448/scalarmult_test.go
@@ -1,16 +1,18 @@
 package edwards448
 
-import "testing"
+import (
+	"testing"
+)
 
 func TestScalarBaseMult(t *testing.T) {
 	t.Run("positive", func(t *testing.T) {
 		s := new(Scalar)
 		g := NewGeneratorPoint()
 		want := new(Point).Zero()
-		for i := 0; i < 1024; i++ {
+		for i := 0; i <= 1024; i++ {
 			got := new(Point).ScalarBaseMult(s)
 			if got.Equal(want) == 0 {
-				t.Errorf("got %#v, want %#v", got, want)
+				t.Errorf("%d: got %#v, want %#v", i, got.Bytes(), want.Bytes())
 			}
 			s.Add(s, &scOne)
 			want.Add(want, g)
@@ -20,15 +22,22 @@ func TestScalarBaseMult(t *testing.T) {
 	t.Run("negative", func(t *testing.T) {
 		s := new(Scalar)
 		g := NewGeneratorPoint()
-		g.Negate(g)
 		want := new(Point).Zero()
-		for i := 0; i < 1024; i++ {
+		for i := 0; i <= 1024; i++ {
 			got := new(Point).ScalarBaseMult(s)
 			if got.Equal(want) == 0 {
-				t.Errorf("got %#v, want %#v", got, want)
+				t.Errorf("%d, got %#v, want %#v", -i, got.Bytes(), want.Bytes())
 			}
 			s.Add(s, &scMinusOne)
-			want.Add(want, g)
+			want.Sub(want, g)
 		}
 	})
+}
+
+func BenchmarkScalarBaseMult(b *testing.B) {
+	s := new(Scalar).Set(&scOne)
+	var p Point
+	for i := 0; i < b.N; i++ {
+		p.ScalarBaseMult(s)
+	}
 }