```
name              old time/op  new time/op  delta
ScalarBaseMult-8   180µs ± 3%   112µs ± 3%  -37.68%  (p=0.000 n=20+19)
```

```
name              old time/op  new time/op  delta
KeyGeneration-8    407µs ± 8%   137µs ± 3%  -66.21%  (p=0.000 n=20+20)
NewKeyFromSeed-8   404µs ± 2%   138µs ± 5%  -65.76%  (p=0.000 n=18+18)
Sign-8             422µs ± 6%   167µs ±21%  -60.52%  (p=0.000 n=20+20)
Verify-8           832µs ± 9%   547µs ± 8%  -34.33%  (p=0.000 n=19+18)
```